delete files related to hetrogenuous graph and move them to IDG1KG repository
